+++ packages/demo/src/configuration/blocks/hardcoded/blockNames.ts
