+++ packages/demo/src/configuration/blocks/hardcoded/effects/blackAndWhiteAndBlurBlock.ts
