+++ packages/demo/src/configuration/blocks/hardcoded/effects/blackAndWhiteBlock.fragment.glsl
