+++ packages/demo/src/configuration/blocks/hardcoded/effects/blackAndWhiteBlock.ts
