+++ packages/demo/src/configuration/blocks/hardcoded/effects/blurBlock.deserializer.ts
