+++ packages/demo/src/configuration/blocks/hardcoded/effects/blurBlock.serializer.ts
