+++ packages/demo/src/configuration/blocks/hardcoded/effects/compositionBlock.ts
