+++ packages/demo/src/configuration/blocks/hardcoded/effects/contrastBlock.ts
