+++ packages/demo/src/configuration/blocks/hardcoded/effects/directionalBlurBlock.serializer.ts
