+++ packages/demo/src/configuration/blocks/hardcoded/effects/directionalBlurBlock.ts
