+++ packages/demo/src/configuration/blocks/hardcoded/effects/exposureBlock.ts
