+++ packages/demo/src/configuration/blocks/hardcoded/effects/frameBlock.ts
