+++ packages/demo/src/configuration/blocks/hardcoded/effects/greenScreenBlock.ts
