+++ packages/demo/src/configuration/blocks/hardcoded/effects/kaleidoscopeBlock.ts
