+++ packages/demo/src/configuration/blocks/hardcoded/effects/maskBlock.ts
