+++ packages/demo/src/configuration/blocks/hardcoded/effects/pixelateBlock.fragment.glsl
