+++ packages/demo/src/configuration/blocks/hardcoded/effects/pixelateBlock.ts
