+++ packages/demo/src/configuration/blocks/hardcoded/effects/posterizeBlock.ts
