+++ packages/demo/src/configuration/blocks/hardcoded/effects/sketchBlock.fragment.glsl
