+++ packages/demo/src/configuration/blocks/hardcoded/effects/sketchBlock.ts
