+++ packages/demo/src/configuration/blocks/hardcoded/effects/softThresholdBlock.ts
