+++ packages/demo/src/configuration/blocks/hardcoded/effects/spritesheetBlock.fragment.glsl
