+++ packages/demo/src/configuration/blocks/hardcoded/effects/spritesheetBlock.ts
