+++ packages/demo/src/configuration/blocks/hardcoded/effects/tintBlock.ts
