+++ packages/demo/src/configuration/blocks/hardcoded/effects/vhsGlitchBlock.ts
