+++ packages/demo/src/configuration/blocks/hardcoded/generators/auroraBlock.fragment.glsl
