+++ packages/demo/src/configuration/blocks/hardcoded/generators/fireworksBlock.fragment.glsl
