+++ packages/demo/src/configuration/blocks/hardcoded/generators/fireworksBlock.ts
@@ -4,95 +4,96 @@ import {
     type SmartFilter,
     type IDisableableBlock,
     type RuntimeData,
+    createStrongRef,
     DisableableShaderBinding,
+    DisableableShaderBlock,
 } from "@babylonjs/smart-filters";
-import { ConnectionPointType, createStrongRef, DisableableShaderBlock } from "@babylonjs/smart-filters";
-import { shaderProgram, uniforms } from "../generators/neonHeartBlock.shader";
+import { ConnectionPointType } from "@babylonjs/smart-filters";
 import { BlockNames } from "../blockNames";
-import { Color3 } from "@babylonjs/core/Maths/math.color";
+import { shaderProgram, uniforms } from "./fireworksBlock.shader";
 
 /**
- * The shader bindings for the Neon Heart block.
+ * The shader bindings for the Fireworks block.
  */
-export class NeonHeartShaderBinding extends DisableableShaderBinding {
+export class FireworksShaderBinding extends DisableableShaderBinding {
     private readonly _inputTexture: RuntimeData<ConnectionPointType.Texture>;
     private readonly _time: RuntimeData<ConnectionPointType.Float>;
-    private readonly _color1: RuntimeData<ConnectionPointType.Color3>;
-    private readonly _color2: RuntimeData<ConnectionPointType.Color3>;
+    private readonly _fireworks: RuntimeData<ConnectionPointType.Float>;
+    private readonly _fireworkSparks: RuntimeData<ConnectionPointType.Float>;
 
     /**
-     * Creates a new shader binding instance for the Neon Heart block.
+     * Creates a new shader binding instance for the Fireworks block.
      * @param parentBlock - The parent block
      * @param inputTexture - The input texture
      * @param time - The time passed since the start of the effect
-     * @param color1 - The first color of the heart
-     * @param color2 - The second color of the heart
+     * @param fireworks - The number of fireworks to display
+     * @param fireworkSparks - The number of sparks per firework
      */
     constructor(
         parentBlock: IDisableableBlock,
         inputTexture: RuntimeData<ConnectionPointType.Texture>,
         time: RuntimeData<ConnectionPointType.Float>,
-        color1: RuntimeData<ConnectionPointType.Color3>,
-        color2: RuntimeData<ConnectionPointType.Color3>
+        fireworks: RuntimeData<ConnectionPointType.Float>,
+        fireworkSparks: RuntimeData<ConnectionPointType.Float>
     ) {
         super(parentBlock);
         this._inputTexture = inputTexture;
         this._time = time;
-        this._color1 = color1;
-        this._color2 = color2;
+        this._fireworks = fireworks;
+        this._fireworkSparks = fireworkSparks;
     }
 
     /**
      * Binds all the required data to the shader when rendering.
      * @param effect - defines the effect to bind the data to
-     * @param _width - defines the width of the output. Passed in
-     * @param _height - defines the height of the output
+     * @param _width - defines the width of the output. Passed in by ShaderRuntime.
+     * @param _height - defines the height of the output. Passed in by ShaderRuntime.
      */
     public override bind(effect: Effect, _width: number, _height: number): void {
         super.bind(effect);
         effect.setTexture(this.getRemappedName(uniforms.input), this._inputTexture.value);
-        effect.setFloat(this.getRemappedName(uniforms.time), this._time.value);
         effect.setFloat(this.getRemappedName(uniforms.aspectRatio), _width / _height);
-        effect.setColor3(this.getRemappedName(uniforms.color1), this._color1.value);
-        effect.setColor3(this.getRemappedName(uniforms.color2), this._color2.value);
+        effect.setFloat(this.getRemappedName(uniforms.time), this._time.value);
+        effect.setFloat(this.getRemappedName(uniforms.fireworks), this._fireworks.value);
+        effect.setFloat(this.getRemappedName(uniforms.fireworkSparks), this._fireworkSparks.value);
     }
 }
 
 /**
- * A procedural block that draws an animated neon heart around the center of the input texture.
+ * A block that adds a fireworks effect overlaid on the input texture.
  */
-export class NeonHeartBlock extends DisableableShaderBlock {
+export class FireworksBlock extends DisableableShaderBlock {
     /**
      * The class name of the block.
      */
-    public static override ClassName = BlockNames.neonHeart;
+    public static override ClassName = BlockNames.fireworks;
 
     /**
-     * The texture connection point.
+     * The input texture connection point
      */
     public readonly input = this._registerInput("input", ConnectionPointType.Texture);
 
     /**
-     * The time connection point.
+     * The time connection point to animate the effect.
      */
     public readonly time = this._registerOptionalInput("time", ConnectionPointType.Float, createStrongRef(0.3));
 
     /**
-     * The color1 connection point, controlling the first color of the heart.
+     * The number of fireworks at a time, as a connection point.
      */
-    public readonly color1 = this._registerOptionalInput(
-        "color1",
-        ConnectionPointType.Color3,
-        createStrongRef(new Color3(1.0, 0.05, 0.3))
+    public readonly fireworks = this._registerOptionalInput(
+        "fireworks",
+        ConnectionPointType.Float,
+        createStrongRef(8.0)
     );
 
     /**
-     * The color2 connection point, controlling the second color of the heart.
+     * The number of sparks per firework, as a connection point.
      */
-    public readonly color2 = this._registerOptionalInput(
-        "color2",
-        ConnectionPointType.Color3,
-        createStrongRef(new Color3(0.1, 0.4, 1.0))
+    public readonly fireworkSparks = this._registerOptionalInput(
+        "fireworkSparks",
+        ConnectionPointType.Float,
+        createStrongRef(128.0)
     );
 
     /**
@@ -106,7 +107,7 @@ export class NeonHeartBlock extends DisableableShaderBlock {
      * @param name - The friendly name of the block
      */
     constructor(smartFilter: SmartFilter, name: string) {
-        super(smartFilter, name);
+        super(smartFilter, name, true);
     }
 
     /**
@@ -116,9 +117,9 @@ export class NeonHeartBlock extends DisableableShaderBlock {
     public getShaderBinding(): DisableableShaderBinding {
         const input = this._confirmRuntimeDataSupplied(this.input);
         const time = this.time.runtimeData;
-        const color1 = this.color1.runtimeData;
-        const color2 = this.color2.runtimeData;
+        const fireworks = this.fireworks.runtimeData;
+        const fireworkSparks = this.fireworkSparks.runtimeData;
 
-        return new NeonHeartShaderBinding(this, input, time, color1, color2);
+        return new FireworksShaderBinding(this, input, time, fireworks, fireworkSparks);
     }
 }
