+++ packages/demo/src/configuration/blocks/hardcoded/generators/heartsBlock.fragment.glsl
