+++ packages/demo/src/configuration/blocks/hardcoded/generators/heartsBlock.ts
@@ -1,99 +1,84 @@
 // For demo and non-commercial usage only
 import type { Effect } from "@babylonjs/core/Materials/effect";
+import type { SmartFilter, IDisableableBlock, RuntimeData } from "@babylonjs/smart-filters";
 import {
-    type SmartFilter,
-    type IDisableableBlock,
-    type RuntimeData,
+    ConnectionPointType,
     createStrongRef,
     DisableableShaderBinding,
     DisableableShaderBlock,
 } from "@babylonjs/smart-filters";
-import { ConnectionPointType } from "@babylonjs/smart-filters";
+import { shaderProgram, uniforms } from "./heartsBlock.shader";
 import { BlockNames } from "../blockNames";
-import { shaderProgram, uniforms } from "../generators/fireworksBlock.shader";
+import { Color3 } from "@babylonjs/core/Maths/math.color";
 
 /**
- * The shader bindings for the Fireworks block.
+ * The shader bindings for the Hearts block.
  */
-export class FireworksShaderBinding extends DisableableShaderBinding {
+export class HeartsShaderBinding extends DisableableShaderBinding {
     private readonly _inputTexture: RuntimeData<ConnectionPointType.Texture>;
     private readonly _time: RuntimeData<ConnectionPointType.Float>;
-    private readonly _fireworks: RuntimeData<ConnectionPointType.Float>;
-    private readonly _fireworkSparks: RuntimeData<ConnectionPointType.Float>;
+    private readonly _tint: RuntimeData<ConnectionPointType.Color3>;
 
     /**
-     * Creates a new shader binding instance for the Fireworks block.
+     * Creates a new shader binding instance for the hearts block.
      * @param parentBlock - The parent block
      * @param inputTexture - The input texture
      * @param time - The time passed since the start of the effect
-     * @param fireworks - The number of fireworks to display
-     * @param fireworkSparks - The number of sparks per firework
+     * @param tint - The color tint of the hearts and vignette
      */
     constructor(
         parentBlock: IDisableableBlock,
         inputTexture: RuntimeData<ConnectionPointType.Texture>,
         time: RuntimeData<ConnectionPointType.Float>,
-        fireworks: RuntimeData<ConnectionPointType.Float>,
-        fireworkSparks: RuntimeData<ConnectionPointType.Float>
+        tint: RuntimeData<ConnectionPointType.Color3>
     ) {
         super(parentBlock);
         this._inputTexture = inputTexture;
         this._time = time;
-        this._fireworks = fireworks;
-        this._fireworkSparks = fireworkSparks;
+        this._tint = tint;
     }
 
     /**
      * Binds all the required data to the shader when rendering.
      * @param effect - defines the effect to bind the data to
-     * @param _width - defines the width of the output. Passed in by ShaderRuntime.
-     * @param _height - defines the height of the output. Passed in by ShaderRuntime.
+     * @param _width - defines the width of the output. Passed in
+     * @param _height - defines the height of the output
      */
     public override bind(effect: Effect, _width: number, _height: number): void {
         super.bind(effect);
         effect.setTexture(this.getRemappedName(uniforms.input), this._inputTexture.value);
-        effect.setFloat(this.getRemappedName(uniforms.aspectRatio), _width / _height);
         effect.setFloat(this.getRemappedName(uniforms.time), this._time.value);
-        effect.setFloat(this.getRemappedName(uniforms.fireworks), this._fireworks.value);
-        effect.setFloat(this.getRemappedName(uniforms.fireworkSparks), this._fireworkSparks.value);
+        effect.setFloat(this.getRemappedName(uniforms.aspectRatio), _width / _height);
+        effect.setColor3(this.getRemappedName(uniforms.tint), this._tint.value);
     }
 }
 
 /**
- * A block that adds a fireworks effect overlaid on the input texture.
+ * A procedural block that renders a heart-emitting tunnel on the input texture.
  */
-export class FireworksBlock extends DisableableShaderBlock {
+export class HeartsBlock extends DisableableShaderBlock {
     /**
      * The class name of the block.
      */
-    public static override ClassName = BlockNames.fireworks;
+    public static override ClassName = BlockNames.hearts;
 
     /**
-     * The input texture connection point
+     * The texture connection point.
      */
     public readonly input = this._registerInput("input", ConnectionPointType.Texture);
 
     /**
-     * The time connection point to animate the effect.
+     * The time connection point.
      */
     public readonly time = this._registerOptionalInput("time", ConnectionPointType.Float, createStrongRef(0.3));
 
     /**
-     * The number of fireworks at a time, as a connection point.
+     * The tint connection point, controlling the color of the hearts and vignette.
      */
-    public readonly fireworks = this._registerOptionalInput(
-        "fireworks",
-        ConnectionPointType.Float,
-        createStrongRef(8.0)
-    );
-
-    /**
-     * The number of sparks per firework, as a connection point.
-     */
-    public readonly fireworkSparks = this._registerOptionalInput(
-        "fireworkSparks",
-        ConnectionPointType.Float,
-        createStrongRef(128.0)
+    public readonly tint = this._registerOptionalInput(
+        "tint",
+        ConnectionPointType.Color3,
+        createStrongRef(new Color3(0.9, 0.5, 0.7))
     );
 
     /**
@@ -107,7 +92,7 @@ export class FireworksBlock extends DisableableShaderBlock {
      * @param name - The friendly name of the block
      */
     constructor(smartFilter: SmartFilter, name: string) {
-        super(smartFilter, name, true);
+        super(smartFilter, name);
     }
 
     /**
@@ -117,9 +102,8 @@ export class FireworksBlock extends DisableableShaderBlock {
     public getShaderBinding(): DisableableShaderBinding {
         const input = this._confirmRuntimeDataSupplied(this.input);
         const time = this.time.runtimeData;
-        const fireworks = this.fireworks.runtimeData;
-        const fireworkSparks = this.fireworkSparks.runtimeData;
+        const tint = this.tint.runtimeData;
 
-        return new FireworksShaderBinding(this, input, time, fireworks, fireworkSparks);
+        return new HeartsShaderBinding(this, input, time, tint);
     }
 }
