+++ packages/demo/src/configuration/blocks/hardcoded/generators/neonHeartBlock.fragment.glsl
