+++ packages/demo/src/configuration/blocks/hardcoded/generators/neonHeartBlock.ts
@@ -7,64 +7,93 @@ import {
     DisableableShaderBinding,
 } from "@babylonjs/smart-filters";
 import { ConnectionPointType, createStrongRef, DisableableShaderBlock } from "@babylonjs/smart-filters";
+import { shaderProgram, uniforms } from "./neonHeartBlock.shader";
 import { BlockNames } from "../blockNames";
-import { shaderProgram, uniforms } from "../generators/tunnelBlock.shader";
+import { Color3 } from "@babylonjs/core/Maths/math.color";
 
 /**
- * The shader bindings for the Tunnel block.
+ * The shader bindings for the Neon Heart block.
  */
-export class TunnelShaderBinding extends DisableableShaderBinding {
+export class NeonHeartShaderBinding extends DisableableShaderBinding {
     private readonly _inputTexture: RuntimeData<ConnectionPointType.Texture>;
     private readonly _time: RuntimeData<ConnectionPointType.Float>;
+    private readonly _color1: RuntimeData<ConnectionPointType.Color3>;
+    private readonly _color2: RuntimeData<ConnectionPointType.Color3>;
 
     /**
-     * Creates a new shader binding instance for the Tunnel block.
+     * Creates a new shader binding instance for the Neon Heart block.
      * @param parentBlock - The parent block
      * @param inputTexture - The input texture
      * @param time - The time passed since the start of the effect
+     * @param color1 - The first color of the heart
+     * @param color2 - The second color of the heart
      */
     constructor(
         parentBlock: IDisableableBlock,
         inputTexture: RuntimeData<ConnectionPointType.Texture>,
-        time: RuntimeData<ConnectionPointType.Float>
+        time: RuntimeData<ConnectionPointType.Float>,
+        color1: RuntimeData<ConnectionPointType.Color3>,
+        color2: RuntimeData<ConnectionPointType.Color3>
     ) {
         super(parentBlock);
         this._inputTexture = inputTexture;
         this._time = time;
+        this._color1 = color1;
+        this._color2 = color2;
     }
 
     /**
      * Binds all the required data to the shader when rendering.
      * @param effect - defines the effect to bind the data to
-     * @param _width - defines the width of the output. Passed in by ShaderRuntime.
-     * @param _height - defines the height of the output. Passed in by ShaderRuntime.
+     * @param _width - defines the width of the output. Passed in
+     * @param _height - defines the height of the output
      */
     public override bind(effect: Effect, _width: number, _height: number): void {
         super.bind(effect);
         effect.setTexture(this.getRemappedName(uniforms.input), this._inputTexture.value);
-        effect.setFloat2(this.getRemappedName(uniforms.resolution), _width, _height);
         effect.setFloat(this.getRemappedName(uniforms.time), this._time.value);
+        effect.setFloat(this.getRemappedName(uniforms.aspectRatio), _width / _height);
+        effect.setColor3(this.getRemappedName(uniforms.color1), this._color1.value);
+        effect.setColor3(this.getRemappedName(uniforms.color2), this._color2.value);
     }
 }
 
 /**
- * A shader block that creates a tunnel-like background effect.
+ * A procedural block that draws an animated neon heart around the center of the input texture.
  */
-export class TunnelBlock extends DisableableShaderBlock {
+export class NeonHeartBlock extends DisableableShaderBlock {
     /**
      * The class name of the block.
      */
-    public static override ClassName = BlockNames.tunnel;
+    public static override ClassName = BlockNames.neonHeart;
 
     /**
-     * The fallback texture to use if block is disabled.
+     * The texture connection point.
      */
-    public readonly fallback = this._registerInput("fallback", ConnectionPointType.Texture);
+    public readonly input = this._registerInput("input", ConnectionPointType.Texture);
 
     /**
      * The time connection point.
      */
     public readonly time = this._registerOptionalInput("time", ConnectionPointType.Float, createStrongRef(0.3));
+
+    /**
+     * The color1 connection point, controlling the first color of the heart.
+     */
+    public readonly color1 = this._registerOptionalInput(
+        "color1",
+        ConnectionPointType.Color3,
+        createStrongRef(new Color3(1.0, 0.05, 0.3))
+    );
+
+    /**
+     * The color2 connection point, controlling the second color of the heart.
+     */
+    public readonly color2 = this._registerOptionalInput(
+        "color2",
+        ConnectionPointType.Color3,
+        createStrongRef(new Color3(0.1, 0.4, 1.0))
+    );
 
     /**
      * The shader program (vertex and fragment code) to use to render the block
@@ -85,9 +114,11 @@ export class TunnelBlock extends DisableableShaderBlock {
      * @returns The class instance that binds the data to the effect
      */
     public getShaderBinding(): DisableableShaderBinding {
-        const input = this._confirmRuntimeDataSupplied(this.fallback);
+        const input = this._confirmRuntimeDataSupplied(this.input);
         const time = this.time.runtimeData;
+        const color1 = this.color1.runtimeData;
+        const color2 = this.color2.runtimeData;
 
-        return new TunnelShaderBinding(this, input, time);
+        return new NeonHeartShaderBinding(this, input, time, color1, color2);
     }
 }
