+++ packages/demo/src/configuration/blocks/hardcoded/generators/particleBlock.ts
