+++ packages/demo/src/configuration/blocks/hardcoded/generators/starryPlanesBlock.fragment.glsl
