+++ packages/demo/src/configuration/blocks/hardcoded/generators/starryPlanesBlock.ts
@@ -1,41 +1,36 @@
 // For demo and non-commercial usage only
 import type { Effect } from "@babylonjs/core/Materials/effect";
-import type { SmartFilter, IDisableableBlock, RuntimeData } from "@babylonjs/smart-filters";
 import {
-    ConnectionPointType,
-    createStrongRef,
+    type SmartFilter,
+    type IDisableableBlock,
+    type RuntimeData,
     DisableableShaderBinding,
-    DisableableShaderBlock,
 } from "@babylonjs/smart-filters";
-import { shaderProgram, uniforms } from "../generators/heartsBlock.shader";
+import { ConnectionPointType, createStrongRef, DisableableShaderBlock } from "@babylonjs/smart-filters";
 import { BlockNames } from "../blockNames";
-import { Color3 } from "@babylonjs/core/Maths/math.color";
+import { shaderProgram, uniforms } from "./starryPlanesBlock.shader";
 
 /**
- * The shader bindings for the Hearts block.
+ * The shader bindings for the StarryPlanes block.
  */
-export class HeartsShaderBinding extends DisableableShaderBinding {
+export class StarryPlanesShaderBinding extends DisableableShaderBinding {
     private readonly _inputTexture: RuntimeData<ConnectionPointType.Texture>;
     private readonly _time: RuntimeData<ConnectionPointType.Float>;
-    private readonly _tint: RuntimeData<ConnectionPointType.Color3>;
 
     /**
-     * Creates a new shader binding instance for the hearts block.
+     * Creates a new shader binding instance for the StarryPlanes block.
      * @param parentBlock - The parent block
      * @param inputTexture - The input texture
      * @param time - The time passed since the start of the effect
-     * @param tint - The color tint of the hearts and vignette
      */
     constructor(
         parentBlock: IDisableableBlock,
         inputTexture: RuntimeData<ConnectionPointType.Texture>,
-        time: RuntimeData<ConnectionPointType.Float>,
-        tint: RuntimeData<ConnectionPointType.Color3>
+        time: RuntimeData<ConnectionPointType.Float>
     ) {
         super(parentBlock);
         this._inputTexture = inputTexture;
         this._time = time;
-        this._tint = tint;
     }
 
     /**
@@ -48,38 +43,28 @@ export class HeartsShaderBinding extends DisableableShaderBinding {
         super.bind(effect);
         effect.setTexture(this.getRemappedName(uniforms.input), this._inputTexture.value);
         effect.setFloat(this.getRemappedName(uniforms.time), this._time.value);
-        effect.setFloat(this.getRemappedName(uniforms.aspectRatio), _width / _height);
-        effect.setColor3(this.getRemappedName(uniforms.tint), this._tint.value);
+        effect.setFloat2(this.getRemappedName(uniforms.resolution), _width, _height);
     }
 }
 
 /**
- * A procedural block that renders a heart-emitting tunnel on the input texture.
+ * A shader block that renders a procedural starry background effect.
  */
-export class HeartsBlock extends DisableableShaderBlock {
+export class StarryPlanesBlock extends DisableableShaderBlock {
     /**
      * The class name of the block.
      */
-    public static override ClassName = BlockNames.hearts;
+    public static override ClassName = BlockNames.starryPlanes;
 
     /**
-     * The texture connection point.
+     * The fallback texture connection point, in the event that the effect is disabled.
      */
-    public readonly input = this._registerInput("input", ConnectionPointType.Texture);
+    public readonly fallback = this._registerInput("fallback", ConnectionPointType.Texture);
 
     /**
      * The time connection point.
      */
     public readonly time = this._registerOptionalInput("time", ConnectionPointType.Float, createStrongRef(0.3));
-
-    /**
-     * The tint connection point, controlling the color of the hearts and vignette.
-     */
-    public readonly tint = this._registerOptionalInput(
-        "tint",
-        ConnectionPointType.Color3,
-        createStrongRef(new Color3(0.9, 0.5, 0.7))
-    );
 
     /**
      * The shader program (vertex and fragment code) to use to render the block
@@ -100,10 +85,9 @@ export class HeartsBlock extends DisableableShaderBlock {
      * @returns The class instance that binds the data to the effect
      */
     public getShaderBinding(): DisableableShaderBinding {
-        const input = this._confirmRuntimeDataSupplied(this.input);
+        const input = this._confirmRuntimeDataSupplied(this.fallback);
         const time = this.time.runtimeData;
-        const tint = this.tint.runtimeData;
 
-        return new HeartsShaderBinding(this, input, time, tint);
+        return new StarryPlanesShaderBinding(this, input, time);
     }
 }
