+++ packages/demo/src/configuration/blocks/hardcoded/inputs/webCamInputBlock.ts
