+++ packages/demo/src/configuration/blocks/hardcoded/inputs/webCamRuntime.ts
