+++ packages/demo/src/configuration/blocks/hardcoded/inputs/webCamSession.ts
