+++ packages/demo/src/configuration/blocks/hardcoded/transitions/glitchBlock.ts
