+++ packages/demo/src/configuration/blocks/hardcoded/transitions/tileBlock.serializer.ts
