+++ packages/demo/src/configuration/blocks/hardcoded/transitions/wipeBlock.deserializer.ts
