+++ packages/demo/src/configuration/blocks/hardcoded/transitions/wipeBlock.serializer.ts
