+++ packages/demo/src/configuration/blocks/hardcoded/utility/premultiplyAlphaBlock.ts
